seperated files into slow and fast
