--- python-performance-optimization-guide/step-2.py
@@ -1,17 +1,0 @@
-import timeit
-
-numbers_list = list(range(100_000))
-
-target = 99_999
-
-list_time = timeit.timeit(lambda: target in numbers_list, number=1000)
-
-print(target in numbers_list)
-print(f"list membership: {list_time:.8f} seconds")
-
-numbers_set = set(numbers_list)
-
-set_time = timeit.timeit(lambda: target in numbers_set, number=1000)
-
-print(target in numbers_set)
-print(f"set membership:  {set_time:.8f} seconds")
