--- python-performance-optimization-guide/step-3.py
@@ -1,30 +1,0 @@
-import timeit
-
-words = ["apple", "banana", "apple", "cherry", "banana", "apple"] * 1000
-
-
-def manual_count(items):
-    counts = {}
-    for item in items:
-        if item in counts:
-            counts[item] += 1
-        else:
-            counts[item] = 1
-    return counts
-
-
-print(manual_count(words))
-
-manual_time = timeit.timeit(lambda: manual_count(words), number=1000)
-print(f"manual loop: {manual_time:.4f} seconds")
-
-from collections import Counter
-
-def counter_count(items):
-    return Counter(items)
-
-
-print(counter_count(words))
-
-counter_time = timeit.timeit(lambda: counter_count(words), number=1000)
-print(f"collections.Counter: {counter_time:.4f} seconds")
